Se hace la ventana de terminos de las soluciones con campos dinámicos
Former-commit-id: 66b1dda2127d975452cd208c63c8bc3c6e600f23

+++ recovery-coreextension-plugin/src/main/java/es/pfsgroup/plugin/recovery/mejoras/acuerdos/MEJAcuerdoManager.java
@@ -609,10 +609,21 @@ public class MEJAcuerdoManager implements MEJAcuerdoApi {
 	@BusinessOperationDefinition(BO_ACUERDO_MGR_DELETE_VALORES_TERMINO)
 	@Transactional(readOnly = false)    
     public void deleteAllValoresTermino(TerminoAcuerdo ta) {
-    	for (ValoresCamposTermino valor : ta.getValoresCampos()) {
-			genericDao.deleteById(ValoresCamposTermino.class, valor.getId());
+		if (ta.getValoresCampos()!=null) {
+	    	for (ValoresCamposTermino valor : ta.getValoresCampos()) {
+				genericDao.deleteById(ValoresCamposTermino.class, valor.getId());
+			}
 		}
     }
+	
+	@Transactional(readOnly = false)
+	public void saveAllValoresTermino(TerminoAcuerdo ta) {
+		if (ta.getValoresCampos()!=null) {
+			for (ValoresCamposTermino valor: ta.getValoresCampos()) {
+				genericDao.save(ValoresCamposTermino.class, valor);
+			}
+		}
+	}
     
 	/**
      * @param to TerminoOperaciones
